Q Learning Agent training code, all data, and png plots
Code includes classes: Game, QLearningAgent, RandomAgent, OptimalAgent
It is easy to create a QLearningAgent with any parameters, opponent, and storage
directory. The QLearningAgent is saved using pickle to easily retrieve
it. Training accuracy data is also saved. Plotting and saving this
accuracy data can be done with either methods or a global function.
Training an agent does not train it from scratch, it picks up where it
left off from before.

The code currently trains the QLearningAgent against the optimal one,
itself, and the random one. In all cases, it converges on the optimal
result. It takes significantly longer for the random one though.