Restructed code. Added new experiment of varying environment.
RESTRUCTURING:
The original main.py file had both the Game and Agent classes.
Now, those are placed into 2 modules names "game" and "agents".
The Q-learning agent now has a new method to set the opponent.
Before it was fixed at the start. The agent also has a new parameter,
"reset", which allows you to make a new object instead of loading from
the file if it exists. The argument for "self.action()" is now an object
of type "Game" instead of list. This is more intuitive and easier to
work with for certain Agents such as the non optimal ones (need xor
sum). Other small bits were improved such as puttin in "Return"
description for docstrings. A bug with the Game.finished() method was
fixed (before it returned if the game was not finished).

ADDITIONS:
Added ReverseOptimalAgent, RandomReverseOptimalAgent, and tested and
analyzed how the Q-learning agent learns in 3 different environments.
Each environment has 2 agents, the optimal and either Random,
ReverseOptimalAgent, or RandomReverseOptimalAgent. The percent of
optimal agents were varied from 0, 1, 2 .. 100. It turns out that
playing with the 2nd "mallicious" agent actually is detrimental and it
takes longer to converge. See plots and comments for more details.

TO DO:
The "Varied Education Level" file has a lot of copy paste that
should be replaced with functions. Also, more trials should be run.